// src/generated/types/UserTypeDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type UserTypeDto = {
  /**
   * @type string
   */
  name: string
  /**
   * @type string | undefined
   */
  description?: string | undefined
  /**
   * @default true
   * @type boolean | undefined
   */
  isDeletable?: boolean | undefined
  /**
   * @type string
   */
  id: string
}