// src/generated/types/UpdateUserTypeDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type UpdateUserTypeDto = {
  /**
   * @type string | undefined
   */
  name?: string | undefined
  /**
   * @type string | undefined
   */
  description?: string | undefined
  /**
   * @default true
   * @type boolean | undefined
   */
  isDeletable?: boolean | undefined
}