// src/generated/types/CreateUserTypeDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { GenericAssignPrivilegeDto } from './GenericAssignPrivilegeDto.ts'

export type CreateUserTypeDto = {
  /**
   * @type string
   */
  name: string
  /**
   * @type string | undefined
   */
  description?: string | undefined
  /**
   * @default true
   * @type boolean | undefined
   */
  isDeletable?: boolean | undefined
  /**
   * @type array
   */
  privilege_assignments: GenericAssignPrivilegeDto[]
}